Move src/lib.rs tests to own file
